+++ src/__tests/utils/index.ts
@@ -3,7 +3,7 @@
 import { Readable } from 'stream'
 import { EventEmitter } from 'events'
 import * as path from 'path'
-import Logger from '../../logger'
+import Logger from '../../utils/logger'
 import { LoggerFactory } from '../../typed'
 
 export function radeStreamAll(stream: Readable, callback: (buffer: Buffer) => any) {
