+++ src/request.ts
@@ -4,8 +4,8 @@ import * as url from 'url'
 import * as http from 'http'
 import * as https from 'https'
 import { EventEmitter } from 'events'
-import CertManager from './cert'
-import * as resources from './res'
+import CertManager from './utils/cert'
+import * as resources from './utils/res'
 import { Request, Response, Logger } from './typed'
 import { emitterPromisify } from './utils/promisify'
