+++ src/utils/res.ts
@@ -4,7 +4,7 @@ import * as fs from 'fs'
 import * as path from 'path'
 
 const cache = {}
-const resdir = '../resources/html'
+const resdir = '../../resources/html'
 
 export function get(filename: string): string {
   let content = cache[filename]
